Change ViewProjectsShowRightHookListener
The listener class is changed in order to allow multiple usage of the
:view_projects_show_right_hook.

Without this change another plugin will override the feature what is
provided over the hook listener in this plugin.

Now another plugin can inherit from the listener class and use super
in order to extend the feature.